requirements.txt
awscli==1.16.124
boto3==1.9.114
botocore==1.12.114
colorama==0.3.9
Django==2.1.4
django-admin-list-filter-dropdown==1.0.2
django-admin-sortable2==0.7
django-cors-headers==2.4.0
django-grappelli==2.12.1
django-modeladmin-reorder==0.3.1
django-s3direct==1.1.3
django-smart-selects==1.5.4
django-storages==1.7.1
django-timezone-field==3.0
django-versatileimagefield==1.10
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docutils==0.14
jmespath==0.9.4
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.7
PyYAML==3.13
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
sorl-thumbnail==12.5.0
urllib3==1.24.1
var-dump==1.2